Fix reset button

+++ Reset.java
@@ -13,34 +13,15 @@ import java.io.IOException;
  */
 public class Reset extends Buttons
 {
-    public Reset() 
-    {
-        setImage(new GreenfootImage("Reset Highscore", 30, Color.BLUE, null));
-        File scores = new File(".scores");
-        try{
-            if(scores.exists()==false){
-                scores.createNewFile();
-                PrintWriter out = new PrintWriter(new FileWriter(scores, false));
-                out.print(0);
-                out.close();
-            }
-        }
-        catch(IOException e){
-            System.out.println("COULD NOT LOG!!");
-        }
+    private boolean applet()  
+    {  
+        return null != System.getSecurityManager();  
     }
 
-    public void act() 
+    public Reset() 
     {
-        setImage(new GreenfootImage("Reset Highscore", 30, Color.BLUE, null));
-
-        if (Greenfoot.getMouseInfo() != null){
-            if (Greenfoot.getMouseInfo().getX() >= 360 && Greenfoot.getMouseInfo().getX() <= 540 &&
-            Greenfoot.getMouseInfo().getY() >= 665 && Greenfoot.getMouseInfo().getY() <= 685) {
-                setImage(new GreenfootImage("Reset Highscore", 30, Color.RED, null));
-            }
-        }
-        if (Greenfoot.mouseClicked(this)) {
+        if(!applet()){
+            setImage(new GreenfootImage("Reset Highscore", 30, Color.BLUE, null));
             File scores = new File(".scores");
             try{
                 if(scores.exists()==false){
@@ -49,16 +30,44 @@ public class Reset extends Buttons
                     out.print(0);
                     out.close();
                 }
-                else{
-                    PrintWriter out = new PrintWriter(new FileWriter(scores, false));
-                    out.print(0);
-                    out.close();
-                }
             }
             catch(IOException e){
                 System.out.println("COULD NOT LOG!!");
             }
-            Greenfoot.setWorld(new Intro());
+        }
+    }
+
+    public void act() 
+    {
+        if(!applet()){
+            setImage(new GreenfootImage("Reset Highscore", 30, Color.BLUE, null));
+
+            if (Greenfoot.getMouseInfo() != null){
+                if (Greenfoot.getMouseInfo().getX() >= 360 && Greenfoot.getMouseInfo().getX() <= 540 &&
+                Greenfoot.getMouseInfo().getY() >= 665 && Greenfoot.getMouseInfo().getY() <= 685) {
+                    setImage(new GreenfootImage("Reset Highscore", 30, Color.RED, null));
+                }
+            }
+            if (Greenfoot.mouseClicked(this)) {
+                File scores = new File(".scores");
+                try{
+                    if(scores.exists()==false){
+                        scores.createNewFile();
+                        PrintWriter out = new PrintWriter(new FileWriter(scores, false));
+                        out.print(0);
+                        out.close();
+                    }
+                    else{
+                        PrintWriter out = new PrintWriter(new FileWriter(scores, false));
+                        out.print(0);
+                        out.close();
+                    }
+                }
+                catch(IOException e){
+                    System.out.println("COULD NOT LOG!!");
+                }
+                Greenfoot.setWorld(new Intro());
+            }
         }
     }
 }
